🔥 Remove retired wizard components and session bridges
Edition edit/add modals now use a slim per-edition form (PricingForm-based);
class-level listing fields are edited only in the status page's Book details
tab. Delete NewNFTBook, RegisterISCN, MintNFT, EditISCNMetadataForm/Modal,
ISCNInfoCard, the sessionStorage upload/descriptionFull bridges, and their
orphaned i18n keys.
