+++ i18n/locales/en.json
@@ -302,14 +302,11 @@
   "copy.description_template": "Copy #{copyNumber} of {nftName}",
   "copy.name_template": "{nftName} #{copyNumber}",
   "edit_iscn_modal.cancel": "Cancel",
-  "edit_iscn.update_failed": "An error occurred while updating metadata",
-  "edit_iscn.update_success": "Metadata updated successfully",
   "error_connect_wallet_failed": "Unable to connect to the wallet",
   "error.fetch_classid_failed": "Some errors occurred, please try again.",
   "error.fetch_nft_metadata_failed": "This ID is invalid",
   "error.nft_id_reserved_for_author": "#0 is the first edition, and we’ll reserve it for you.",
   "errors.add_moderator_wallet": "Please press Add button to add moderator wallet",
-  "errors.add_notification_email": "Please press Add button to add notification email",
   "errors.contact_support": "Contact Support",
   "errors.edition_not_found": "Edition not found",
   "errors.file_size_limit": "File exceeds 1MB limit",
@@ -354,7 +351,6 @@
   "form.disable_download": "Disable Download",
   "form.disable_file_download": "Disable File Download",
   "form.drm_options": "Unique title page",
-  "form.edit_iscn_metadata": "Update metadata",
   "form.edition": "Edition",
   "form.email_notifications": "Email to receive sales notifications",
   "form.enable_advanced_settings": "Please enable this feature in the advanced settings",
@@ -508,7 +504,6 @@
   "nft_book_form.accept_tipping_tooltip": "Readers can provide additional support beyond the book price",
   "nft_book_form.add_custom_message_page": "Add the custom page",
   "nft_book_form.add_edition": "Add Edition",
-  "nft_book_form.advanced_settings": "Advanced Settings",
   "nft_book_form.author_message": "Author's message ✍️:",
   "nft_book_form.auto_delivery": "Auto delivery",
   "nft_book_form.auto_delivery_memo": "Preset message (optional)",
@@ -590,7 +585,6 @@
   "nft.minting_restock": "Printing new stock",
   "nft.minting_restock_in_progress": "Printing {count} copies...please wait",
   "notifications.file_upload_complete": "Please do not close this window until upload is complete.",
-  "notifications.iscn_registration_complete": "Please do not close this window until metadata registration is complete.",
   "pages.add": "Add",
   "pages.add_coupon": "Add Coupon",
   "pages.add_coupon_btn": "Add Coupon",
@@ -735,17 +729,8 @@
   "preview_book.supported_formats": "Supports EPUB, PDF, and image formats",
   "preview_book.title": "Preview Book",
   "prices.standard_edition": "eBook version",
-  "publish_button.metadata": "Register Metadata",
-  "publish_button.mint_nft": "Mint",
   "publish_button.publish_now": "Publish",
-  "publish_button.upload_files": "Start Upload",
   "publish_nft_book.page_title": "Publish New Work",
-  "publish_steps.already_have_iscn": "Already have ISCN ID? Click ",
-  "publish_steps.continue_last_register": "Continue previous registration:",
-  "publish_steps.enter_iscn_id": "Enter ISCN ID",
-  "publish_steps.here": "here",
-  "publish_steps.mint_nft": "Mint First Edition",
-  "publish_steps.new_book_publish": "List on 3ookStore",
   "publish_progress.creating_listing": "Creating book listing...",
   "publish_progress.creating_nft_class": "Creating NFT book...",
   "publish_progress.minting_nft": "Minting first edition...",
@@ -755,8 +740,6 @@
   "publish_review.free": "Free",
   "publish_review.metadata_title": "Book Details",
   "publish_review.pricing_title": "Pricing & Editions",
-  "publish_steps.register_iscn": "Register Metadata",
-  "publish_steps.upload_files": "Upload Files",
   "publish_wizard.back": "Back",
   "publish_wizard.next": "Next",
   "publish_wizard.publishing_do_not_close": "Please keep this page open until publishing completes",
