+++ i18n/locales/zh-TW.json
@@ -302,14 +302,11 @@
   "copy.description_template": "{nftName} 的第 {copyNumber} 份副本",
   "copy.name_template": "{nftName} #{copyNumber}",
   "edit_iscn_modal.cancel": "取消",
-  "edit_iscn.update_failed": "更新元資料時發生錯誤",
-  "edit_iscn.update_success": "成功更新元資料",
   "error_connect_wallet_failed": "無法連結錢包",
   "error.fetch_classid_failed": "發生一些錯誤，請重試。",
   "error.fetch_nft_metadata_failed": "該 ID 無效",
   "error.nft_id_reserved_for_author": "#0 為首刷版，我們會將它保留給你",
   "errors.add_moderator_wallet": "請按「新增」按鈕來新增審核錢包",
-  "errors.add_notification_email": "請按「新增」按鈕來新增通知電子郵件",
   "errors.contact_support": "聯絡客服",
   "errors.edition_not_found": "找不到版本",
   "errors.file_size_limit": "檔案超過 1MB 限制",
@@ -354,7 +351,6 @@
   "form.disable_download": "停用下載",
   "form.disable_file_download": "停用檔案下載",
   "form.drm_options": "獨特扉頁",
-  "form.edit_iscn_metadata": "更新元資料",
   "form.edition": "版本",
   "form.email_notifications": "欲收到銷售通知的電郵",
   "form.enable_advanced_settings": "請在進階設定中啟用此功能",
@@ -508,7 +504,6 @@
   "nft_book_form.accept_tipping_tooltip": "讀者可在書價外作進一步支持",
   "nft_book_form.add_custom_message_page": "加入訂製扉頁",
   "nft_book_form.add_edition": "新增版本",
-  "nft_book_form.advanced_settings": "進階",
   "nft_book_form.author_message": "作者留言 ✍️：",
   "nft_book_form.auto_delivery": "自動發書",
   "nft_book_form.auto_delivery_memo": "預設留言（可留空）",
@@ -590,7 +585,6 @@
   "nft.minting_restock": "印鑄新庫存",
   "nft.minting_restock_in_progress": "正在加印 {count} 本...請稍候",
   "notifications.file_upload_complete": "請勿關閉此視窗，直到上傳完成。",
-  "notifications.iscn_registration_complete": "請勿關閉此視窗，直至元資料登記完成。",
   "pages.add": "新增",
   "pages.add_coupon": "新增優惠券",
   "pages.add_coupon_btn": "新增優惠券",
@@ -735,17 +729,8 @@
   "preview_book.supported_formats": "支援 EPUB、PDF 及圖片格式",
   "preview_book.title": "預覽書籍",
   "prices.standard_edition": "電子版",
-  "publish_button.metadata": "登記元資料",
-  "publish_button.mint_nft": "鑄造",
   "publish_button.publish_now": "發佈",
-  "publish_button.upload_files": "開始上傳",
   "publish_nft_book.page_title": "出版新作",
-  "publish_steps.already_have_iscn": "已經有 ISCN ID 了？ 按",
-  "publish_steps.continue_last_register": "繼續上次的註冊:",
-  "publish_steps.enter_iscn_id": "輸入 ISCN ID",
-  "publish_steps.here": "這裡",
-  "publish_steps.mint_nft": "印鑄首刷",
-  "publish_steps.new_book_publish": "上架 3ookStore",
   "publish_progress.creating_listing": "正在建立販售資訊...",
   "publish_progress.creating_nft_class": "正在建立 NFT 書籍...",
   "publish_progress.minting_nft": "正在印鑄首刷...",
@@ -755,8 +740,6 @@
   "publish_review.free": "免費",
   "publish_review.metadata_title": "書籍資料",
   "publish_review.pricing_title": "定價與版本",
-  "publish_steps.register_iscn": "登記元資料",
-  "publish_steps.upload_files": "上傳文本",
   "publish_wizard.back": "上一步",
   "publish_wizard.next": "下一步",
   "publish_wizard.publishing_do_not_close": "發佈完成前請勿關閉此頁面",
